requirements.txt
django==3.2.9
djangorestframework==3.12.4
djongo==1.3.6
matplotlib==3.5.1



